requirements.txt
requests
python-dotenv
discord
responses
imgkit
wkhtmltopdf
